class 05/runner-methods/pyproject.toml
[project]
name = "runner-methods"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.3",
]
